commented out options in next.config,js created a GetGithubData express-helper to make routes cleaner
