+++ next.config.js
@@ -4,10 +4,10 @@ const withCSS = require('@zeit/next-css');
 
 const nextConfig = withCSS({
 	// target: 'server',
-	transformManifest: manifest => ['/'].concat(manifest), // add the homepage to the cache
+	// transformManifest: manifest => ['/'].concat(manifest), // add the homepage to the cache
 	// Trying to set NODE_ENV=production when running yarn dev causes a build-time error so we
 	// turn on the SW in dev mode so that we can actually test it
-	generateInDevMode: true,
+	// generateInDevMode: true,
 	// workboxOpts: {
 	// 	swDest: 'static/service-worker.js',
 	// 	runtimeCaching: [
